skills/nano-banana-manager/requirements.txt
# nano-banana-manager dependencies
fastmcp>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
Pillow>=10.0.0
